requirements.txt
aniso8601==9.0.1
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
flask-restx==1.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
marshmallow==3.19.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
